requirements.txt
gradio
huggingface-hub
Pillow
python-dotenv